move copy specialization into sys::unix module
