+++ library/std/src/sys/unix/mod.rs
@@ -51,6 +51,8 @@ pub mod fd;
 pub mod fs;
 pub mod futex;
 pub mod io;
+#[cfg(any(target_os = "linux", target_os = "android"))]
+pub mod kernel_copy;
 #[cfg(target_os = "l4re")]
 mod l4re;
 pub mod memchr;
